adds dynamic config file functionality and improves test coverage
Before, the config.json file location was hardcoded in the code which made it difficult to test from both packages and additionally
it meant that the user was forced into following the same config.json file structure. This has been changed so you can dynamically provide
the location of the config.json file as an argument to the application. It has some additional checking that identifies if there's a correct path etc

Additionally the unit test coverage has now increase